fix(hub): fix IE issues by keying tokens w/ appId
IE fails w/ permissionError when trying to use a removed iframe as a key
during cleanup. This causes the cleanup to fail.

This commit introduces an appId to token map, that allows cleanup to
happen regardless of window state.

BREAKING CHANGE: hub.forgetApp API signature now requires appId as well
as target window.